all tpetra changes in one commit
continued debugging

continued debugging

+++ src/solvers/SolverContinuation.hh
@@ -17,8 +17,8 @@ the initial guess to solve a harder problem.  As each successive problem is
 solved, the continuation parameter is changed closer and closer to the true
 value.
 
-Few if any PKs support this method currently -- it requires the PK to provide more
-interface about how to update the continuation parameter.
+Few if any PKs support this method currently -- it requires the PK to provide
+more interface about how to update the continuation parameter.
 
 .. _solver-continuation-spec:
 .. admonition:: solver-continuation-spec
@@ -77,7 +77,7 @@ class SolverContinuation : public Solver<Vector, VectorSpace> {
     tol_ = tol;
     solver_->set_tolerance(tol);
   }
-  void set_pc_lag(int pc_lag) { solver_->set_pc_lag(pc_lag); }
+  void set_pc_lag(double pc_lag) { solver_->set_pc_lag(pc_lag); }
 
   // access
   double tolerance() { return tol_; }
@@ -108,8 +108,8 @@ class SolverContinuation : public Solver<Vector, VectorSpace> {
 
 
 /* ******************************************************************
-* Public Init method.
-****************************************************************** */
+ * Public Init method.
+ ****************************************************************** */
 template <class Vector, class VectorSpace>
 void
 SolverContinuation<Vector, VectorSpace>::Init(const Teuchos::RCP<SolverFnBase<Vector>>& fn,
@@ -122,8 +122,8 @@ SolverContinuation<Vector, VectorSpace>::Init(const Teuchos::RCP<SolverFnBase<Ve
 
 
 /* ******************************************************************
-* Initialization of the NKA solver.
-****************************************************************** */
+ * Initialization of the NKA solver.
+ ****************************************************************** */
 template <class Vector, class VectorSpace>
 void
 SolverContinuation<Vector, VectorSpace>::Init_()
@@ -140,8 +140,8 @@ SolverContinuation<Vector, VectorSpace>::Init_()
 
 
 /* ******************************************************************
-* The body of NKA solver
-****************************************************************** */
+ * The body of NKA solver
+ ****************************************************************** */
 template <class Vector, class VectorSpace>
 int
 SolverContinuation<Vector, VectorSpace>::Solve_(const Teuchos::RCP<Vector>& u)
